+++ ovsnl/conntrack_common.go
@@ -1,0 +1,61 @@
+// Copyright 2017 DigitalOcean.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package ovsnl
+
+import (
+	"net"
+)
+
+// ConntrackEntry represents a single connection tracking entry from the kernel.
+type ConntrackEntry struct {
+	Protocol   string // "tcp", "udp", "icmp" etc.
+	OrigSrc    net.IP
+	OrigDst    net.IP
+	OrigSPort  uint16
+	OrigDPort  uint16
+	ReplySrc   net.IP
+	ReplyDst   net.IP
+	ReplySPort uint16
+	ReplyDPort uint16
+	Zone       uint16
+	Mark       uint32
+	State      string
+}
+
+// ZoneStats holds statistics for a zone
+type ZoneStats struct {
+	TotalCount int
+	Entries    []ConntrackEntry // Only populated if TotalCount > threshold
+}
+
+// ConntrackPerformanceStats represents aggregated performance counters from all CPUs
+type ConntrackPerformanceStats struct {
+	TotalFound         uint32
+	TotalInvalid       uint32
+	TotalIgnore        uint32
+	TotalInsert        uint32
+	TotalInsertFailed  uint32
+	TotalDrop          uint32
+	TotalEarlyDrop     uint32
+	TotalError         uint32
+	TotalSearchRestart uint32
+	CPUs               int
+}
+
+// ZmKey is a compact key for (zone,mark)
+type ZmKey struct {
+	Zone uint16
+	Mark uint32
+}
