+++ ovsnl/conntrack_stub.go
@@ -18,51 +18,8 @@ package ovsnl
 
 import (
 	"context"
-	"fmt"
-	"net"
+	"errors"
 )
-
-// ConntrackEntry represents a single connection tracking entry from the kernel.
-type ConntrackEntry struct {
-	Protocol   string // "tcp", "udp", "icmp" etc.
-	OrigSrc    net.IP
-	OrigDst    net.IP
-	OrigSPort  uint16
-	OrigDPort  uint16
-	ReplySrc   net.IP
-	ReplyDst   net.IP
-	ReplySPort uint16
-	ReplyDPort uint16
-	Zone       uint16
-	Mark       uint32
-	State      string
-}
-
-// ZoneStats holds statistics for a zone
-type ZoneStats struct {
-	TotalCount int
-	Entries    []ConntrackEntry // Only populated if TotalCount > threshold
-}
-
-// ConntrackPerformanceStats represents aggregated performance counters from all CPUs
-type ConntrackPerformanceStats struct {
-	TotalFound         uint32
-	TotalInvalid       uint32
-	TotalIgnore        uint32
-	TotalInsert        uint32
-	TotalInsertFailed  uint32
-	TotalDrop          uint32
-	TotalEarlyDrop     uint32
-	TotalError         uint32
-	TotalSearchRestart uint32
-	CPUs               int
-}
-
-// ZmKey is a compact key for (zone,mark)
-type ZmKey struct {
-	Zone uint16
-	Mark uint32
-}
 
 // ZoneMarkAggregator keeps live counts (zone -> mark -> count).
 type ZoneMarkAggregator struct {
